Extract signup into single page to allow for better error handling
